requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0